refactor: date field
